Move headers and source files out of the types/ directory again.
