+++ dynd/include/kernels/apply_jit_kernel.hpp
@@ -3,8 +3,8 @@
 #include <dynd/kernels/base_kernel.hpp>
 #include <array_functions.hpp>
 #include <utility_functions.hpp>
-#include <types/type_functions.hpp>
 
+#include "type_functions.hpp"
 #include "array_conversions.hpp"
 
 namespace pydynd {
