+++ dynd/include/type_functions.hpp
