+++ dynd/src/array_from_py.cpp
@@ -24,8 +24,8 @@
 #include "array_functions.hpp"
 #include "array_conversions.hpp"
 #include "numpy_interop.hpp"
-#include "types/type_functions.hpp"
-#include "types/type_deduction.hpp"
+#include "type_functions.hpp"
+#include "type_deduction.hpp"
 #include "types/pyobject_type.hpp"
 #include "utility_functions.hpp"
